requirements.txt
# AI Agent Orchestration Platform - Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Agent Orchestration & LLM Integration
langgraph==0.0.66
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.0

# Database & Storage
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0  # PostgreSQL async driver
redis==5.0.1
psycopg2-binary==2.9.9

# Vector Database
chromadb==0.4.18
# Alternative: pinecone-client==2.2.4

# Async & Task Queue
celery==5.3.4
kombu==5.3.4
flower==2.0.1  # Celery monitoring

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data Processing & Validation
pandas==2.1.4
numpy==1.25.2
jsonschema==4.20.0

# Monitoring & Observability  
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
structlog==23.2.0

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2

# Utilities
tenacity==8.2.3  # Retry logic
schedule==1.2.0  # Job scheduling
jinja2==3.1.2   # Template engine
click==8.1.7     # CLI interface
rich==13.7.0     # Rich terminal output

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI
factory-boy==3.3.0  # Test data factories

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.10
mkdocstrings[python]==0.24.0